+++ ICLIO/examples/dcclearisochannels.cpp
@@ -11,7 +11,7 @@ int main(int n, char **ppc){
   cammap m;
   
   std::vector<DCDevice> devs = DCGrabber::getDeviceList();
-  printf("found %d cameras \n",devs.size());
+  printf("found %d cameras \n",(unsigned int)devs.size());
   
   
   for(unsigned int i=0;i<devs.size();i++){
