+++ ICLIO/examples/dcdeviceinfo.cpp
@@ -6,7 +6,7 @@ using namespace icl;
 
 int main(int n, char **ppc){
   std::vector<DCDevice> devs = DCGrabber::getDeviceList();
-  printf("found %d cameras \n",devs.size());
+  printf("found %d cameras \n",(unsigned int)devs.size());
 
   
   for(unsigned int i=0;i<devs.size();i++){
